+++ STM32/Src/trx_manager.c
@@ -306,7 +306,7 @@ void TRX_setFrequency(uint32_t _freq, VFO *vfo)
 
 void TRX_setTXFrequencyFloat(float64_t _freq, VFO *vfo)
 {
-	TRX_setFrequency(_freq, vfo);
+	TRX_setFrequency((uint32_t)_freq, vfo);
 	TRX_freq_phrase_tx = getTXPhraseFromFrequency(_freq);
 	FPGA_NeedSendParams = true;
 }
